tasklib/requirements.txt
argparse
daemonize
pyyaml